move FilterArray to utils, add retrieving values
